refactor(server): split Telegram adapter into 4 layers
