+++ apps/server/src/channel/index.ts
@@ -1,4 +1,4 @@
 export * from "./discord/index.js";
-export * from "./telegram.js";
+export * from "./telegram/index.js";
 export * from "./github/index.js";
 export * from "./websocket.js";
