+++ apps/server/src/channel/telegram/index.ts
@@ -1,0 +1,1 @@
+export { TelegramAdapter } from "./surface";
