+++ apps/server/src/channel/telegram/types.ts
@@ -1,0 +1,35 @@
+export interface TelegramUser {
+  id: number;
+  is_bot: boolean;
+  first_name: string;
+  username?: string;
+}
+
+export interface TelegramChat {
+  id: number;
+  type: "private" | "group" | "supergroup" | "channel";
+  title?: string;
+  first_name?: string;
+  username?: string;
+}
+
+export interface TelegramMessage {
+  message_id: number;
+  from?: TelegramUser;
+  chat: TelegramChat;
+  date: number;
+  text?: string;
+}
+
+export interface TelegramUpdate {
+  update_id: number;
+  message?: TelegramMessage;
+}
+
+export interface TelegramResponse<T> {
+  ok: boolean;
+  result: T;
+  description?: string;
+  error_code?: number;
+  parameters?: { retry_after?: number };
+}
